generalization for and parity with non_blocking
