+++ src/non_blocking/mod.rs
@@ -1,15 +1,17 @@
 //use core::fmt::Result;
 //use core::fmt::Write;
 
-use embedded_hal::digital;
-use embedded_hal::digital::OutputPin;
 use embedded_hal_async::delay::DelayNs;
-use embedded_hal_async::i2c::I2c;
 
 pub mod bus;
-use bus::{DataBus, EightBitBus, FourBitBus, I2CBus};
+use bus::DataBus;
 
+use crate::charset::CharsetWithFallback;
+use crate::display_size::DisplaySize;
 pub use crate::error;
+use crate::memory_map::DisplayMemoryMap;
+use crate::sealed::Internal;
+use crate::setup::non_blocking::DisplayOptions;
 use error::Result;
 
 pub use crate::entry_mode;
@@ -20,8 +22,10 @@ pub use crate::display_mode;
 
 pub use display_mode::DisplayMode;
 
-pub struct HD44780<B: DataBus> {
+pub struct HD44780<B: DataBus, M: DisplayMemoryMap, C: CharsetWithFallback> {
 	bus: B,
+	memory_map: M,
+	charset: C,
 	entry_mode: EntryMode,
 	display_mode: DisplayMode,
 }
@@ -31,151 +35,47 @@ pub use crate::CursorBlink;
 pub use crate::Direction;
 pub use crate::Display;
 
-impl<
-		RS: OutputPin<Error = E> + 'static,
-		EN: OutputPin<Error = E> + 'static,
-		D0: OutputPin<Error = E> + 'static,
-		D1: OutputPin<Error = E> + 'static,
-		D2: OutputPin<Error = E> + 'static,
-		D3: OutputPin<Error = E> + 'static,
-		D4: OutputPin<Error = E> + 'static,
-		D5: OutputPin<Error = E> + 'static,
-		D6: OutputPin<Error = E> + 'static,
-		D7: OutputPin<Error = E> + 'static,
-		E: digital::Error,
-	> HD44780<EightBitBus<RS, EN, D0, D1, D2, D3, D4, D5, D6, D7>>
-{
-	/// Create an instance of a `HD44780` from 8 data pins, a register select
-	/// pin, an enable pin and a struct implementing the delay trait.
-	/// - The delay instance is used to sleep between commands to
-	/// ensure the `HD44780` has enough time to process commands.
-	/// - The eight db0..db7 pins are used to send and recieve with
-	///  the `HD44780`.
-	/// - The register select pin is used to tell the `HD44780`
-	/// if incoming data is a command or data.
-	/// - The enable pin is used to tell the `HD44780` that there
-	/// is data on the 8 data pins and that it should read them in.
-	///
-	pub async fn new_8bit<'a, D: DelayNs>(
-		rs: RS,
-		en: EN,
-		d0: D0,
-		d1: D1,
-		d2: D2,
-		d3: D3,
-		d4: D4,
-		d5: D5,
-		d6: D6,
-		d7: D7,
-		delay: &'a mut D,
-	) -> Result<HD44780<EightBitBus<RS, EN, D0, D1, D2, D3, D4, D5, D6, D7>>, E> {
-		let mut hd = HD44780 {
-			bus: EightBitBus::from_pins(rs, en, d0, d1, d2, d3, d4, d5, d6, d7),
-			entry_mode: EntryMode::default(),
-			display_mode: DisplayMode::default(),
-		};
+use self::error::Error;
 
-		hd.init_8bit(delay).await?;
-
-		return Ok(hd);
-	}
-}
-
-impl<
-		RS: OutputPin<Error = E> + 'static,
-		EN: OutputPin<Error = E> + 'static,
-		D4: OutputPin<Error = E> + 'static,
-		D5: OutputPin<Error = E> + 'static,
-		D6: OutputPin<Error = E> + 'static,
-		D7: OutputPin<Error = E> + 'static,
-		E: digital::Error,
-	> HD44780<FourBitBus<RS, EN, D4, D5, D6, D7>>
-{
-	/// Create an instance of a `HD44780` from 4 data pins, a register select
-	/// pin, an enable pin and a struct implementing the delay trait.
-	/// - The delay instance is used to sleep between commands to
-	/// ensure the `HD44780` has enough time to process commands.
-	/// - The four db0..db3 pins are used to send and recieve with
-	///  the `HD44780`.
-	/// - The register select pin is used to tell the `HD44780`
-	/// if incoming data is a command or data.
-	/// - The enable pin is used to tell the `HD44780` that there
-	/// is data on the 4 data pins and that it should read them in.
-	///
-	/// This mode operates differently than 8 bit mode by using 4 less
-	/// pins for data, which is nice on devices with less I/O although
-	/// the I/O takes a 'bit' longer
-	///
-	/// Instead of commands being sent byte by byte each command is
-	/// broken up into it's upper and lower nibbles (4 bits) before
-	/// being sent over the data bus
-	///
-	pub async fn new_4bit<'a, D: DelayNs>(
-		rs: RS,
-		en: EN,
-		d4: D4,
-		d5: D5,
-		d6: D6,
-		d7: D7,
-		delay: &'a mut D,
-	) -> Result<HD44780<FourBitBus<RS, EN, D4, D5, D6, D7>>, E> {
-		let mut hd = HD44780 {
-			bus: FourBitBus::from_pins(rs, en, d4, d5, d6, d7),
-			entry_mode: EntryMode::default(),
-			display_mode: DisplayMode::default(),
-		};
-
-		hd.init_4bit(delay).await?;
-
-		return Ok(hd);
-	}
-}
-
-impl<I2C: I2c + 'static> HD44780<I2CBus<I2C>> {
-	/// Create an instance of a `HD44780` from an I2C bus and a struct implementing the delay trait.
-	/// - The delay instance is used to sleep between commands to
-	/// ensure the `HD44780` has enough time to process commands.
-	/// - The I2C bus is used to send and receive with
-	///  the `HD44780`.
-	///
-	/// This mode operates differently than 8 bit  and 4 bit modes by using
-	/// an I2C expander, which is nice on devices with less I/O although
-	/// the I/O takes a 'bit' longer, because of transmission delays
-	///
-	/// Instead of commands being sent byte by byte each command is
-	/// broken up into it's upper and lower nibbles (4 bits) before
-	/// being sent over the data bus
-	///
-	pub async fn new_i2c<'a, D: DelayNs>(
-		i2c: I2C,
-		address: u8,
-		delay: &'a mut D,
-	) -> Result<HD44780<I2CBus<I2C>>, I2C::Error> {
-		let mut hd = HD44780 {
-			bus: I2CBus::new(i2c, address),
-			entry_mode: EntryMode::default(),
-			display_mode: DisplayMode::default(),
-		};
-
-		hd.init_4bit(delay).await?;
-
-		return Ok(hd);
-	}
-}
-
-impl<B> HD44780<B>
+impl<B, M, C> HD44780<B, M, C>
 where
 	B: DataBus,
+	M: DisplayMemoryMap,
+	C: CharsetWithFallback,
 {
+	/// Create an instance of a `HD44780` using a struct implementing
+	/// the delay trait.
+	/// The delay instance is used to sleep between commands to
+	/// ensure the `HD44780` has enough time to process commands.
+	///
+	/// If there was an error when setting up the display, the settings
+	/// are returned as a tuple together with the error. This can be used
+	/// to retry on error, or just to get back access to registers or buses.
+	pub async fn new<Opt, D: DelayNs>(
+		options: Opt,
+		delay: &mut D,
+	) -> core::result::Result<Self, (Opt, Error<Opt::IoError>)>
+	where
+		Opt: DisplayOptions<Bus = B, MemoryMap = M, Charset = C>,
+	{
+		options.new_display(delay, Internal).await
+	}
+
+	pub fn destroy(self) -> B {
+		self.bus
+	}
+
+	pub(crate) fn new_raw(bus: B, memory_map: M, charset: C, entry_mode: EntryMode, display_mode: DisplayMode) -> Self {
+		Self { bus, memory_map, charset, entry_mode, display_mode }
+	}
+
 	/// Unshifts the display and sets the cursor position to 0
 	///
 	/// ```rust,ignore
-	/// lcd.reset();
+	/// lcd.reset().await?;
 	/// ```
 	pub async fn reset<'a, D: DelayNs>(&mut self, delay: &'a mut D) -> Result<(), B::Error> {
-		self.write_command(0b0000_0010, delay).await?;
-
-		Ok(())
+		self.write_command(0b0000_0010, delay).await
 	}
 
 	/// Set if the display should be on, if the cursor should be
@@ -200,28 +100,34 @@ where
 	/// Clear the entire display
 	///
 	/// ```rust,ignore
-	/// lcd.clear();
+	/// lcd.clear().await?;
 	/// ```
 	pub async fn clear<'a, D: DelayNs>(&mut self, delay: &'a mut D) -> Result<(), B::Error> {
-		self.write_command(0b0000_0001, delay).await?;
+		self.write_command(0b0000_0001, delay).await
+	}
 
-		Ok(())
+	/// Get the memory map information for this display.
+	pub fn memory_map(&self) -> &M {
+		&self.memory_map
+	}
+
+	/// Get the display size.
+	pub fn display_size(&self) -> DisplaySize {
+		self.memory_map.display_size()
 	}
 
 	/// If enabled, automatically scroll the display when a new
 	/// character is written to the display
 	///
 	/// ```rust,ignore
-	/// lcd.set_autoscroll(true);
+	/// lcd.set_autoscroll(true).await?;
 	/// ```
 	pub async fn set_autoscroll<'a, D: DelayNs>(&mut self, enabled: bool, delay: &'a mut D) -> Result<(), B::Error> {
 		self.entry_mode.shift_mode = enabled.into();
 
 		let cmd = self.entry_mode.as_byte();
 
-		self.write_command(cmd, delay).await?;
-
-		Ok(())
+		self.write_command(cmd, delay).await
 	}
 
 	/// Set if the cursor should be visible
@@ -234,9 +140,7 @@ where
 
 		let cmd = self.display_mode.as_byte();
 
-		self.write_command(cmd, delay).await?;
-
-		Ok(())
+		self.write_command(cmd, delay).await
 	}
 
 	/// Set if the characters on the display should be visible
@@ -245,9 +149,7 @@ where
 
 		let cmd = self.display_mode.as_byte();
 
-		self.write_command(cmd, delay).await?;
-
-		Ok(())
+		self.write_command(cmd, delay).await
 	}
 
 	/// Set if the cursor should blink
@@ -260,19 +162,17 @@ where
 
 		let cmd = self.display_mode.as_byte();
 
-		self.write_command(cmd, delay).await?;
-
-		Ok(())
+		self.write_command(cmd, delay).await
 	}
 
 	/// Set which way the cursor will move when a new character is written
 	///
 	/// ```rust,ignore
 	/// // Move right (Default) when a new character is written
-	/// lcd.set_cursor_mode(CursorMode::Right)
+	/// lcd.set_cursor_mode(CursorMode::Right).await?;
 	///
 	/// // Move left when a new character is written
-	/// lcd.set_cursor_mode(CursorMode::Left)
+	/// lcd.set_cursor_mode(CursorMode::Left).await?;
 	/// ```
 	pub async fn set_cursor_mode<'a, D: DelayNs>(
 		&mut self,
@@ -283,30 +183,41 @@ where
 
 		let cmd = self.entry_mode.as_byte();
 
-		self.write_command(cmd, delay).await?;
-
-		Ok(())
+		self.write_command(cmd, delay).await
 	}
 
 	/// Set the cursor position
 	///
 	/// ```rust,ignore
 	/// // Move to line 2
-	/// lcd.set_cursor_pos(40)
+	/// lcd.set_cursor_pos(40).await?;
 	/// ```
 	pub async fn set_cursor_pos<'a, D: DelayNs>(&mut self, position: u8, delay: &'a mut D) -> Result<(), B::Error> {
 		let lower_7_bits = 0b0111_1111 & position;
 
-		self.write_command(0b1000_0000 | lower_7_bits, delay).await?;
+		self.write_command(0b1000_0000 | lower_7_bits, delay).await
+	}
 
-		Ok(())
+	/// Set the cursor position
+	///
+	/// ```rust,ignore
+	/// // Move to the start of line 3
+	/// lcd.set_cursor_pos_xy(0,2).await?;
+	/// ```
+	pub async fn set_cursor_xy<D: DelayNs>(&mut self, position: (u8, u8), delay: &mut D) -> Result<(), B::Error> {
+		let size = self.display_size().get();
+		let Some(pos) = self.memory_map.address_for_xy(position.0, position.1) else {
+			return Err(Error::Position { position, size });
+		};
+
+		self.write_command(0b1000_0000 | pos, delay).await
 	}
 
 	/// Shift just the cursor to the left or the right
 	///
 	/// ```rust,ignore
-	/// lcd.shift_cursor(Direction::Left);
-	/// lcd.shift_cursor(Direction::Right);
+	/// lcd.shift_cursor(Direction::Left).await?;
+	/// lcd.shift_cursor(Direction::Right).await?;
 	/// ```
 	pub async fn shift_cursor<'a, D: DelayNs>(&mut self, dir: Direction, delay: &'a mut D) -> Result<(), B::Error> {
 		let bits = match dir {
@@ -314,16 +225,14 @@ where
 			Direction::Right => 0b0000_0100,
 		};
 
-		self.write_command(0b0001_0000 | bits | bits, delay).await?;
-
-		Ok(())
+		self.write_command(0b0001_0000 | bits | bits, delay).await
 	}
 
 	/// Shift the entire display to the left or the right
 	///
 	/// ```rust,ignore
-	/// lcd.shift_display(Direction::Left);
-	/// lcd.shift_display(Direction::Right);
+	/// lcd.shift_display(Direction::Left).await?;
+	/// lcd.shift_display(Direction::Right).await?;
 	/// ```
 	pub async fn shift_display<'a, D: DelayNs>(&mut self, dir: Direction, delay: &'a mut D) -> Result<(), B::Error> {
 		let bits = match dir {
@@ -331,9 +240,7 @@ where
 			Direction::Right => 0b0000_0100,
 		};
 
-		self.write_command(0b0001_1000 | bits, delay).await?;
-
-		Ok(())
+		self.write_command(0b0001_1000 | bits, delay).await
 	}
 
 	/// Write a single character to the `HD44780`. This `char` just gets downcast to a `u8`
@@ -342,10 +249,10 @@ where
 	/// See the documentation on that function for more details about compatibility.
 	///
 	/// ```rust,ignore
-	/// lcd.write_char('A', &'a mut DelayUs)?; // prints 'A'
+	/// lcd.write_char('A', &'a mut DelayUs).await?; // prints 'A'
 	/// ```
 	pub async fn write_char<'a, D: DelayNs>(&mut self, data: char, delay: &'a mut D) -> Result<(), B::Error> {
-		self.write_byte(data as u8, delay).await
+		self.write_byte(self.charset.code_from_utf8_with_fallback(data), delay).await
 	}
 
 	async fn write_command<'a, D: DelayNs>(&mut self, cmd: u8, delay: &'a mut D) -> Result<(), B::Error> {
@@ -356,112 +263,25 @@ where
 		Ok(())
 	}
 
-	async fn init_4bit<'a, D: DelayNs>(&mut self, delay: &'a mut D) -> Result<(), B::Error> {
-		// Wait for the LCD to wakeup if it was off
-		delay.delay_ms(15).await;
-
-		// Initialize Lcd in 4-bit mode
-		self.bus.write(0x33, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_ms(5).await;
-
-		// Sets 4-bit operation and enables 5x7 mode for chars
-		self.bus.write(0x32, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		self.bus.write(0x28, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Clear Display
-		self.bus.write(0x0E, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Move the cursor to beginning of first line
-		self.bus.write(0x01, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Set entry mode
-		self.bus.write(self.entry_mode.as_byte(), false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		self.bus.write(0x80, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		Ok(())
-	}
-
-	// Follow the 8-bit setup procedure as specified in the HD44780 datasheet
-	async fn init_8bit<'a, D: DelayNs>(&mut self, delay: &'a mut D) -> Result<(), B::Error> {
-		// Wait for the LCD to wakeup if it was off
-		delay.delay_ms(15).await;
-
-		// Initialize Lcd in 8-bit mode
-		self.bus.write(0b0011_0000, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_ms(5).await;
-
-		// Sets 8-bit operation and enables 5x7 mode for chars
-		self.bus.write(0b0011_1000, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		self.bus.write(0b0000_1110, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Clear Display
-		self.bus.write(0b0000_0001, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Move the cursor to beginning of first line
-		self.bus.write(0b000_0111, false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		// Set entry mode
-		self.bus.write(self.entry_mode.as_byte(), false, delay).await?;
-
-		// Wait for the command to be processed
-		delay.delay_us(100).await;
-
-		Ok(())
-	}
-
 	/// Writes a string to the HD44780. Internally, this just prints the string byte-by-byte, so
 	/// make sure the characters in the string fit in a normal `u8`. See the documentation on
 	/// [write_byte](#method.write_byte) for more details on compatibility.
 	///
 	/// ```rust,ignore
-	/// lcd.write_str("Hello, World!", &'a mut DelayUs)?;
+	/// lcd.write_str("Hello, World!", &'a mut DelayUs).await?;
 	/// ```
 	pub async fn write_str<'a, D: DelayNs>(&mut self, string: &str, delay: &'a mut D) -> Result<(), B::Error> {
-		self.write_bytes(string.as_bytes(), delay).await
+		for ch in string.chars() {
+			self.write_char(ch, delay).await?;
+		}
+		Ok(())
 	}
 
 	/// Writes a sequence of bytes to the HD44780. See the documentation on the
 	/// [write_byte](#method.write_byte) function for more details about compatibility.
 	///
 	/// ```rust,ignore
-	/// lcd.write_bytes(b"Hello, World!", &'a mut DelayUs)?;
+	/// lcd.write_bytes(b"Hello, World!", &'a mut DelayUs).await?;
 	/// ```
 	pub async fn write_bytes<'a, D: DelayNs>(&mut self, string: &[u8], delay: &'a mut D) -> Result<(), B::Error> {
 		for &b in string {
